update docker os_image to ubuntu18 and gcc7
Summary:
This diff updates docker os_image to ubuntu18 and gcc7 in travis settings.
The dependencies that we work with internally (e.g. gcc7, boost and so on) are not present in the ubuntu-16.04. Further, the support for gcc5 is going away. So maintaining this compatibility for these platforms is painful, especially for new projects, such as mvfst.

Reviewed By: simpkins

Differential Revision: D15286181

fbshipit-source-id: eb0da9556cdb17d0aa132ac2aa4c35f9c6eccd97